Revert "Statemind 169 170 alternative 2"
