+++ README.md
@@ -228,10 +228,10 @@ File                                                     |  % Stmts | % Branch |
   FullMath.sol                                           |      100 |    66.67 |      100 |      100 |                |
   OracleLibrary.sol                                      |      100 |    66.67 |      100 |      100 |                |
   TickMath.sol                                           |      100 |    80.43 |      100 |    84.09 |... 63,65,67,73 |
-  TwapGetter.sol                                         |      100 |     62.5 |      100 |    93.33 |             57 |
- contracts\peer-to-peer\wrappers\ERC20\                  |      100 |    79.63 |      100 |    97.75 |                |
-  ERC20Wrapper.sol                                       |      100 |       75 |      100 |    96.88 |             45 |
-  WrappedERC20Impl.sol                                   |      100 |    82.35 |      100 |    98.25 |            151 |
+  TwapGetter.sol                                         |      100 |     62.5 |      100 |    93.33 |             53 |
+ contracts\peer-to-peer\wrappers\ERC20\                  |    97.83 |    80.43 |    92.86 |    96.05 |                |
+  ERC20Wrapper.sol                                       |    93.75 |    77.27 |       80 |    93.75 |          45,84 |
+  WrappedERC20Impl.sol                                   |      100 |    83.33 |      100 |    97.73 |            124 |
  contracts\peer-to-peer\wrappers\ERC721\                 |    98.72 |    85.29 |    94.12 |    96.77 |                |
   ERC721Wrapper.sol                                      |    95.45 |    81.82 |       80 |    92.68 |      47,87,135 |
   WrappedERC721Impl.sol                                  |      100 |    86.96 |      100 |    98.25 |        194,212 |
@@ -245,6 +245,6 @@ File                                                     |  % Stmts | % Branch |
   IFundingPoolImpl.sol                                   |      100 |      100 |      100 |      100 |                |
   ILoanProposalImpl.sol                                  |      100 |      100 |      100 |      100 |                |
 ---------------------------------------------------------|----------|----------|----------|----------|----------------|
-All files                                                |    98.99 |    88.59 |    98.63 |    96.74 |                |
+All files                                                |    98.87 |     88.7 |    98.17 |    96.65 |                |
 ---------------------------------------------------------|----------|----------|----------|----------|----------------|
 ```
