+++ contracts/Errors.sol
@@ -122,5 +122,4 @@ library Errors {
     error TokenAlreadyCountedInWrapper();
     error TokenNotOwnedByWrapper();
     error TokenDoesNotBelongInWrapper(address tokenAddr, uint256 tokenId);
-    error InvalidMintAmount();
 }
