+++ test/peer-to-peer/local-tests.ts
@@ -4,9 +4,8 @@ import { StandardMerkleTree } from '@openzeppelin/merkle-tree'
 import { LenderVaultImpl, MyERC20 } from '../../typechain-types'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
 import { payloadScheme } from './helpers/abi'
-import { setupBorrowerWhitelist, getSlot, findBalanceSlot } from './helpers/misc'
+import { setupBorrowerWhitelist } from './helpers/misc'
 import { HARDHAT_CHAIN_ID_AND_FORKING_CONFIG } from '../../hardhat.config'
-import { erc20 } from '../../typechain-types/@openzeppelin/contracts/token'
 
 // test config vars
 let snapshotId: String // use snapshot id to reset state before each test
@@ -20,8 +19,7 @@ const MAX_UINT256 = ethers.BigNumber.from(2).pow(256).sub(1)
 const ONE_DAY = ethers.BigNumber.from(60 * 60 * 24)
 const ZERO_BYTES32 = ethers.utils.formatBytes32String('')
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
-const SINGLE_WRAPPER_REDEMPTION_FEE = BASE.mul(5).div(1000)
-const SINGLE_WRAPPER_MIN_MINT = ethers.BigNumber.from(10).pow(6)
+const YEAR_IN_SECONDS = 31_536_000
 
 async function generateOffChainQuote({
   lenderVault,
@@ -4531,297 +4529,6 @@ describe('Peer-to-Peer: Local Tests', function () {
 
       expect(await wrappedSingleToken.totalSupply()).to.equal(ONE_WETH.mul(2))
       expect(await wrappedSingleToken.balanceOf(team.address)).to.equal(ONE_WETH.mul(1))
-
-      const allTokensCreated = await erc20Wrapper.allTokensCreated()
-      expect(allTokensCreated[0]).to.be.equal(await erc20Wrapper.tokensCreated(0))
-      expect(allTokensCreated[1]).to.be.equal(await erc20Wrapper.tokensCreated(1))
-      expect(allTokensCreated[2]).to.be.equal(await erc20Wrapper.tokensCreated(2))
-    })
-
-    it('Should handle single underling with negative rebasing correctly', async function () {
-      const { addressRegistry, borrower, team, weth, erc20Wrapper } = await setupTest()
-
-      // mint test tokens
-      await weth.mint(borrower.address, ONE_WETH.mul(1000))
-      await weth.mint(team.address, ONE_WETH.mul(1000))
-
-      // set token wrapper contract in address registry
-      await addressRegistry.connect(team).setWhitelistState([erc20Wrapper.address], 8)
-
-      // approve wrapper for initial mint
-      await weth.connect(borrower).approve(erc20Wrapper.address, MAX_UINT256)
-
-      // create wrapped token
-      const initialMintAmount = ONE_WETH
-      await addressRegistry.connect(borrower).createWrappedTokenForERC20s(
-        [
-          {
-            tokenAddr: weth.address,
-            tokenAmount: initialMintAmount
-          }
-        ],
-        'wethWrapper',
-        'wWeth',
-        ZERO_BYTES32
-      )
-
-      // check new token has been created
-      expect(await erc20Wrapper.numTokensCreated()).to.be.equal(1)
-
-      // get new wrapper token contract
-      const newSingleWrappedTokenAddr = await erc20Wrapper.tokensCreated(0)
-      const wrappedSingleToken = await ethers.getContractAt('WrappedERC20Impl', newSingleWrappedTokenAddr)
-
-      // check received mint amount
-      expect(await wrappedSingleToken.balanceOf(borrower.address)).to.equal(initialMintAmount.sub(SINGLE_WRAPPER_MIN_MINT))
-
-      // approve and mint again
-      const secondMintAmount = ONE_WETH.mul(2)
-      await weth.connect(borrower).approve(wrappedSingleToken.address, MAX_UINT256)
-      await wrappedSingleToken.connect(borrower).mint(borrower.address, secondMintAmount, 0)
-
-      expect(await wrappedSingleToken.totalSupply()).to.equal(initialMintAmount.add(secondMintAmount))
-      expect(await wrappedSingleToken.balanceOf(borrower.address)).to.equal(
-        initialMintAmount.sub(SINGLE_WRAPPER_MIN_MINT).add(secondMintAmount)
-      )
-
-      // automatically find mapping slot
-      const mappingSlot = await findBalanceSlot(weth)
-
-      // calculate balanceOf wrappedSingleToken slot
-      const signerBalanceSlot = getSlot(wrappedSingleToken.address, mappingSlot)
-
-      // set it to the value to mock negative rebasing token
-      const targetBal = ONE_WETH
-      await ethers.provider.send('hardhat_setStorageAt', [
-        weth.address,
-        signerBalanceSlot,
-        ethers.utils.hexlify(ethers.utils.zeroPad(targetBal, 32))
-      ])
-
-      // check that the user balance is equal to the expected value
-      expect(await weth.balanceOf(wrappedSingleToken.address)).to.be.eq(targetBal)
-
-      // check that another mint takes into account downwards rebased wrapper balance
-      await weth.connect(team).approve(wrappedSingleToken.address, MAX_UINT256)
-      const preMintBal = await wrappedSingleToken.balanceOf(team.address)
-      const totalSupply = await wrappedSingleToken.totalSupply()
-      const thirdMintAmount = ONE_WETH.mul(2)
-      await wrappedSingleToken.connect(team).mint(team.address, thirdMintAmount, 0)
-      const postMintBal = await wrappedSingleToken.balanceOf(team.address)
-      const expMintAmount = thirdMintAmount.mul(totalSupply).div(targetBal)
-      expect(expMintAmount).to.be.eq(postMintBal.sub(preMintBal))
-
-      // check redemption
-      const addressRegistryOwner = await addressRegistry.owner()
-      const totalBal = await weth.balanceOf(wrappedSingleToken.address)
-      const preRedemptionBal = await weth.balanceOf(borrower.address)
-      const preRedemptionBalOfAddressRegistry = await weth.balanceOf(addressRegistryOwner)
-      const preRedemptionTotalSupply = await wrappedSingleToken.totalSupply()
-      const redemptionAmount = await wrappedSingleToken.balanceOf(borrower.address)
-      await wrappedSingleToken.connect(borrower).redeem(borrower.address, borrower.address, redemptionAmount)
-      const postRedemptionBal = await weth.balanceOf(borrower.address)
-      const postRedemptionTotalSupply = await wrappedSingleToken.totalSupply()
-      const expReceivedRedemptionAmountWithoutFees = redemptionAmount.mul(totalBal).div(preRedemptionTotalSupply)
-      const redemptionFee = expReceivedRedemptionAmountWithoutFees.mul(SINGLE_WRAPPER_REDEMPTION_FEE).div(BASE)
-      const postRedemptionBalOfAddressRegistry = await weth.balanceOf(addressRegistryOwner)
-      expect(expReceivedRedemptionAmountWithoutFees.sub(redemptionFee)).to.be.eq(postRedemptionBal.sub(preRedemptionBal))
-      expect(postRedemptionBalOfAddressRegistry.sub(preRedemptionBalOfAddressRegistry)).to.be.equal(redemptionFee)
-      expect(preRedemptionTotalSupply.sub(redemptionAmount)).to.be.equal(postRedemptionTotalSupply)
-    })
-
-    it('Should handle single underling with positive rebasing correctly', async function () {
-      const { addressRegistry, borrower, team, weth, erc20Wrapper } = await setupTest()
-
-      // mint test tokens
-      await weth.mint(borrower.address, ONE_WETH.mul(1000))
-      await weth.mint(team.address, ONE_WETH.mul(1000))
-
-      // set token wrapper contract in address registry
-      await addressRegistry.connect(team).setWhitelistState([erc20Wrapper.address], 8)
-
-      // approve wrapper for initial mint
-      await weth.connect(borrower).approve(erc20Wrapper.address, MAX_UINT256)
-
-      // create wrapped token
-      const initialMintAmount = ONE_WETH
-      await addressRegistry.connect(borrower).createWrappedTokenForERC20s(
-        [
-          {
-            tokenAddr: weth.address,
-            tokenAmount: initialMintAmount
-          }
-        ],
-        'wethWrapper',
-        'wWeth',
-        ZERO_BYTES32
-      )
-
-      // check new token has been created
-      expect(await erc20Wrapper.numTokensCreated()).to.be.equal(1)
-
-      // get new wrapper token contract
-      const newSingleWrappedTokenAddr = await erc20Wrapper.tokensCreated(0)
-      const wrappedSingleToken = await ethers.getContractAt('WrappedERC20Impl', newSingleWrappedTokenAddr)
-
-      // check received mint amount
-      expect(await wrappedSingleToken.balanceOf(borrower.address)).to.equal(initialMintAmount.sub(SINGLE_WRAPPER_MIN_MINT))
-
-      // approve and mint again
-      const secondMintAmount = ONE_WETH.mul(2)
-      await weth.connect(borrower).approve(wrappedSingleToken.address, MAX_UINT256)
-      await wrappedSingleToken.connect(borrower).mint(borrower.address, secondMintAmount, 0)
-
-      expect(await wrappedSingleToken.totalSupply()).to.equal(initialMintAmount.add(secondMintAmount))
-      expect(await wrappedSingleToken.balanceOf(borrower.address)).to.equal(
-        initialMintAmount.sub(SINGLE_WRAPPER_MIN_MINT).add(secondMintAmount)
-      )
-
-      // automatically find mapping slot
-      const mappingSlot = await findBalanceSlot(weth)
-
-      // calculate balanceOf wrappedSingleToken slot
-      const signerBalanceSlot = getSlot(wrappedSingleToken.address, mappingSlot)
-
-      // set it to the value to mock negative rebasing token
-      const targetBal = ONE_WETH.mul(1000)
-      await ethers.provider.send('hardhat_setStorageAt', [
-        weth.address,
-        signerBalanceSlot,
-        ethers.utils.hexlify(ethers.utils.zeroPad(targetBal, 32))
-      ])
-
-      // check that the user balance is equal to the expected value
-      expect(await weth.balanceOf(wrappedSingleToken.address)).to.be.eq(targetBal)
-
-      // check that another mint takes into account upwards rebased wrapper balance
-      await weth.connect(team).approve(wrappedSingleToken.address, MAX_UINT256)
-      const preMintWrappedTokenUndBal = await weth.balanceOf(wrappedSingleToken.address)
-      const preMintUserBal = await weth.balanceOf(team.address)
-      const preMintTotalSupply = await wrappedSingleToken.totalSupply()
-      const thirdMintAmount = ONE_WETH.mul(2)
-      await wrappedSingleToken.connect(team).mint(team.address, thirdMintAmount, 0)
-      const postMintWrappedTokenUndBal = await weth.balanceOf(wrappedSingleToken.address)
-      const postMintUserBal = await weth.balanceOf(team.address)
-      const postMintTotalSupply = await wrappedSingleToken.totalSupply()
-      const expMintAmount = thirdMintAmount.mul(preMintTotalSupply).div(preMintWrappedTokenUndBal)
-      expect(thirdMintAmount).to.be.eq(postMintWrappedTokenUndBal.sub(preMintWrappedTokenUndBal))
-      expect(thirdMintAmount).to.be.eq(preMintUserBal.sub(postMintUserBal))
-      expect(expMintAmount).to.be.eq(postMintTotalSupply.sub(preMintTotalSupply))
-
-      // check redemption
-      const addressRegistryOwner = await addressRegistry.owner()
-      const totalBal = await weth.balanceOf(wrappedSingleToken.address)
-      const preRedemptionBal = await weth.balanceOf(borrower.address)
-      const preRedemptionBalOfAddressRegistry = await weth.balanceOf(addressRegistryOwner)
-      const preRedemptionTotalSupply = await wrappedSingleToken.totalSupply()
-      const redemptionAmount = await wrappedSingleToken.balanceOf(borrower.address)
-      await wrappedSingleToken.connect(borrower).redeem(borrower.address, borrower.address, redemptionAmount)
-      const postRedemptionBal = await weth.balanceOf(borrower.address)
-      const postRedemptionTotalSupply = await wrappedSingleToken.totalSupply()
-      const expReceivedRedemptionAmountWithoutFees = redemptionAmount.mul(totalBal).div(preRedemptionTotalSupply)
-      const redemptionFee = expReceivedRedemptionAmountWithoutFees.mul(SINGLE_WRAPPER_REDEMPTION_FEE).div(BASE)
-      const postRedemptionBalOfAddressRegistry = await weth.balanceOf(addressRegistryOwner)
-      expect(expReceivedRedemptionAmountWithoutFees.sub(redemptionFee)).to.be.eq(postRedemptionBal.sub(preRedemptionBal))
-      expect(postRedemptionBalOfAddressRegistry.sub(preRedemptionBalOfAddressRegistry)).to.be.equal(redemptionFee)
-      expect(preRedemptionTotalSupply.sub(redemptionAmount)).to.be.equal(postRedemptionTotalSupply)
-    })
-
-    it('Should handle single underlying with donations correctly', async function () {
-      const { addressRegistry, wrappedERC20Impl, borrower, team, weth, erc20Wrapper } = await setupTest()
-
-      // mint test tokens
-      await weth.mint(borrower.address, ONE_WETH.mul(1000))
-      await weth.mint(team.address, ONE_WETH.mul(1000))
-
-      // set token wrapper contract in address registry
-      await addressRegistry.connect(team).setWhitelistState([erc20Wrapper.address], 8)
-
-      // approve wrapper for initial mint
-      await weth.connect(borrower).approve(erc20Wrapper.address, MAX_UINT256)
-
-      // check revert on too small mint amount
-      await expect(
-        addressRegistry.connect(borrower).createWrappedTokenForERC20s(
-          [
-            {
-              tokenAddr: weth.address,
-              tokenAmount: 1
-            }
-          ],
-          'wethWrapper',
-          'wWeth',
-          ZERO_BYTES32
-        )
-      ).to.be.revertedWithCustomError(wrappedERC20Impl, 'InvalidMintAmount')
-
-      // create wrapped token
-      const initialMintAmount = SINGLE_WRAPPER_MIN_MINT.add(1)
-      await addressRegistry.connect(borrower).createWrappedTokenForERC20s(
-        [
-          {
-            tokenAddr: weth.address,
-            tokenAmount: initialMintAmount
-          }
-        ],
-        'wethWrapper',
-        'wWeth',
-        ZERO_BYTES32
-      )
-
-      // check new token has been created
-      expect(await erc20Wrapper.numTokensCreated()).to.be.equal(1)
-
-      // get new wrapper token contract
-      const newSingleWrappedTokenAddr = await erc20Wrapper.tokensCreated(0)
-      const wrappedSingleToken = await ethers.getContractAt('WrappedERC20Impl', newSingleWrappedTokenAddr)
-
-      // check received mint amount
-      expect(await wrappedSingleToken.balanceOf(borrower.address)).to.equal(initialMintAmount.sub(SINGLE_WRAPPER_MIN_MINT))
-
-      // check expected balances
-      expect(await wrappedSingleToken.totalSupply()).to.equal(initialMintAmount)
-      expect(await weth.balanceOf(wrappedSingleToken.address)).to.equal(initialMintAmount)
-
-      // specify some 2nd mint amount
-      const secondMint = ONE_WETH
-
-      // front-run mint with donation that causes truncation
-      const totalSupply = await wrappedSingleToken.totalSupply()
-      const totalBalance = await weth.balanceOf(wrappedSingleToken.address)
-      const donationToCauseTruncation = secondMint.mul(totalSupply).sub(totalBalance).add(1)
-      await weth.mint(wrappedSingleToken.address, donationToCauseTruncation)
-
-      // check that underlying balance increased by donation amount
-      expect(await weth.balanceOf(wrappedSingleToken.address)).to.be.eq(donationToCauseTruncation.add(initialMintAmount))
-
-      // check that mint reverts
-      await weth.connect(team).approve(wrappedSingleToken.address, MAX_UINT256)
-      await expect(wrappedSingleToken.connect(team).mint(team.address, secondMint, 0)).to.be.revertedWithCustomError(
-        wrappedSingleToken,
-        'InvalidMintAmount'
-      )
-
-      // check redemption
-      const preRedemptionUserTokenBal = await wrappedSingleToken.balanceOf(borrower.address)
-      const preRedemptionUserUndBal = await weth.balanceOf(borrower.address)
-      const preRedemptionWrapperUndBal = await weth.balanceOf(wrappedSingleToken.address)
-      const preRedemptionTokenSupply = await wrappedSingleToken.totalSupply()
-      await wrappedSingleToken.connect(borrower).redeem(borrower.address, borrower.address, preRedemptionUserTokenBal)
-      const postRedemptionUserTokenBal = await wrappedSingleToken.balanceOf(borrower.address)
-      const postRedemptionUserUndBal = await weth.balanceOf(borrower.address)
-      const postRedemptionWrapperUndBal = await weth.balanceOf(wrappedSingleToken.address)
-      const postRedemptionTotalSupply = await wrappedSingleToken.totalSupply()
-      const expReceivedRedemptionAmountWithoutFees = preRedemptionUserTokenBal
-        .mul(preRedemptionWrapperUndBal)
-        .div(preRedemptionTokenSupply)
-      const expRedemptionFee = expReceivedRedemptionAmountWithoutFees.mul(SINGLE_WRAPPER_REDEMPTION_FEE).div(BASE)
-      expect(postRedemptionUserTokenBal).to.be.equal(0)
-      expect(postRedemptionTotalSupply).to.be.equal(SINGLE_WRAPPER_MIN_MINT)
-      expect(postRedemptionWrapperUndBal).to.be.equal(preRedemptionWrapperUndBal.sub(expReceivedRedemptionAmountWithoutFees))
-      expect(postRedemptionUserUndBal.sub(preRedemptionUserUndBal)).to.be.equal(
-        expReceivedRedemptionAmountWithoutFees.sub(expRedemptionFee)
-      )
     })
   })
